// src/app/custom-layout/components/customer-component/customer-details/branches/interfaces/contact.interface.ts
export class Contact {
  id: number;
  imageSrc: string;
  bName?:string;
  addressLine1?:string;
  name: string;
  email: string;
  phone: string;
  company: string;
  notes?: string;
  birthday?: string;
  selected: boolean;
  starred: boolean;
  city?: string;
  country?: string;
  addressLine2?: string;
}